save: Use build tags for deciding default path during compile
Switch from deciding the os specific path at runtime to during compile time.
