+++ pkg/save/utils.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"runtime"
 	"strconv"
 	"strings"
 )
@@ -27,27 +26,6 @@ func SpaceshipParts() []string {
 	list := make([]string, len(spaceshipParts))
 	copy(list, spaceshipParts)
 	return list
-}
-
-func DefaultSaveLocation() (string, error) {
-	home, err := os.UserHomeDir()
-	if err != nil {
-		return "", err
-	}
-
-	var path string
-	switch runtime.GOOS {
-	case "windows":
-		path = saveFolderWindows(home)
-	case "linux":
-		path = filepath.Join(home, "snap/steam/common/.local/share/Steam/steamapps/compatdata/1511460/pfx/drive_c/users/steamuser/")
-		path = saveFolderWindows(path)
-	}
-	if _, err := os.Stat(path); path != "" && !os.IsNotExist(err) {
-		return path, nil
-	} else {
-		return home, nil
-	}
 }
 
 func saveFolderWindows(root string) string {
